projects_with_fastai/TODO_APP/requirements.txt
fastapi 
uvicorn 
SQLAlchemy
passlib
bcrypt
pyjwt
python-multipart